test(cli): extract attribute tradition play tests
Move the attribute purchase, tradition swap, and assignment-review closeout command tests into a focused play suite. The new file owns the player-operation validation/send fake state for BUY_ATTRIBUTE_TREE_NODE, CHANGE_TRADITION, CONSIDER_ASSIGN_ATTRIBUTE, and CONSIDER_ASSIGN_TRADITIONS.

Keep the monolith responsible for read-side traditions, progress-dashboard, notification, and compact-priority projection fixtures. Wire the focused suite into test:cli:play so the extracted coverage remains canonical.

Proof: git diff --check; bun run --cwd packages/cli test -- game.play.attribute-tradition.test.ts; bun run --cwd packages/cli test -- game.play.test.ts -t "attribute|tradition|progress-dashboard|compact priorities"; bun run check:cli; bun run test:cli:play. This is test ownership modularization only; it changes no runtime behavior, direct-control behavior, live-game flow, gameplay send, play-thread workflow, Studio/oRPC transport, or relationship/city-state/suzerain evidence policy.
